implementing touch screen

+++ device/TSC2046/test-TSC2046/test-TSC2046.cpp
@@ -14,8 +14,8 @@ int main()
 	TSC2046 touchScreen(spi0, GPIO6, GPIO7);
 	touchScreen.Initialize();
 	for (;;) {
-		uint16_t x, y;
-		uint8_t z1, z2;
+		int x, y;
+		int z1, z2;
 		bool touched = touchScreen.ReadPosition(&x, &y, &z1, &z2);
 		if (touched) ::printf("%4x %4x %2x %2x %d\n", x, y, z1, z2, (z2 / z1 - 1) * x * 100 / 4096);
 		::sleep_ms(100);
